requirement.txt
python-telegram-bot==20.6
requests==2.31.0
python-dotenv==1.1.0
